Use f64 instead of f32 to prevent shadow issues

+++ examples/chapter5.rs
@@ -6,7 +6,7 @@ fn main() -> Result<(), fmt::Error> {
     let wall_z = 10.0;
     let wall_size = 7.0;
     let canvas_pixels = 100;
-    let pixel_size = wall_size / f32::from(canvas_pixels);
+    let pixel_size = wall_size / f64::from(canvas_pixels);
     let half = wall_size / 2.0;
 
     let mut canvas = Canvas::new(canvas_pixels, canvas_pixels);
@@ -14,10 +14,10 @@ fn main() -> Result<(), fmt::Error> {
     let shape = Sphere::default();
 
     for y in 0..canvas_pixels {
-        let world_y = half - pixel_size * f32::from(y);
+        let world_y = half - pixel_size * f64::from(y);
 
         for x in 0..canvas_pixels {
-            let world_x = -half + pixel_size * f32::from(x);
+            let world_x = -half + pixel_size * f64::from(x);
             let position = Point::new(world_x, world_y, wall_z);
             let ray = Ray::new(ray_origin, (position - ray_origin).normalize());
