+++ src/lib.rs
@@ -27,4 +27,4 @@ pub use crate::render::render;
 pub use crate::vector::{Vector, Vector3, Vector4};
 pub use crate::world::World;
 
-const EPSILON: f32 = 1e-4;
+const EPSILON: f64 = 1e-4;
